+++ gfx/gfx_text_field2.cpp
@@ -323,7 +323,7 @@ namespace gfx {
 		{
 			return -1;
 		}
-		if (line >= mNewLinePositions.size())
+		if (line >= static_cast<int>(mNewLinePositions.size()))
 		{
 			return -2;
 		}
